backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
qdrant-client==1.14.2
openai==1.82.0
pymupdf==1.25.5
beautifulsoup4==4.13.4
python-docx==1.1.2
pydantic-settings==2.9.1
rank-bm25==0.2.2
